requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pydantic-settings
pydantic[email]
bcrypt
python-jose[cryptography]
python-multipart
pypdf
sentence-transformers
faiss-cpu
numpy
aiofiles
groq
